phase2-web/backend/requirements.txt
# Vercel-compatible requirements (pure Python only)
fastapi==0.124.4
uvicorn==0.38.0
sqlmodel==0.0.27
pyjwt==2.10.1
python-dotenv==1.2.1
httpx==0.28.1
sqlalchemy==2.0.45
pydantic==2.12.5
starlette==0.50.0
anyio==4.12.0
typing-extensions==4.15.0
# Pure Python PostgreSQL driver
pg8000==1.31.2
# Pure Python bcrypt alternative
passlib==1.7.4
